2.6 test merge sprint2401 (#686)
* [Improve][API] Unify type system api(data & type)

* [Feature][JDBC] Add Inceptor data type convertor

* Update oracle/db2/mysql varchar/byte type mapping

* [Feature][JDBC] Add Snowflake data type convertor

+++ seatunnel-connectors-v2/connector-dolphindb-wt/src/main/java/org/apache/seatunnel/connectors/dolphindb/catalog/DolphinDBCatalog.java
@@ -167,7 +167,7 @@ public class DolphinDBCatalog implements Catalog {
                                 columnName,
                                 dolphinDBDataTypeConvertor.toSeaTunnelType(
                                         columnName, dataType.name()),
-                                null,
+                                (Long) null,
                                 true,
                                 null,
                                 null);
